🔥(daemon): get rid of the old `daemon` server, as it isn't used anymore
